First Commit
Added some more features.
* Added timer, that speeds up when level is increased
* Whenever a row is completed, it is removed and all the upper rows fall
down one step. The falling happens gradually so the player can see a
cascading fall.
* Score is displayed every time. Score is added whenever a new piece
reaches the bottom. Score is added whenever a row is completed. Added
scores increase in value with level.
* Pieces can be rotated 90 degrees each time around the top left corner
* Pieces can be mirrored upwards or horizontally
* Some optimizations which reduced the amount of code written. Current
piece on the board is now represented using the offsets currentRow,
currentCol.
* Removed the confusing set(x,y,value) and just using the convention of
board[row][col] everywhere.

Programmed the keyboard input for windows. Don't have a mac to test
this.




In this version, 
* added periodically timed 'moveDown's
* bug fixes
   - prints the board at the start of the game
   - Won't move left if it will collide
   - Won't move right if it will collide
* code autoformatted in eclipse